backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Werkzeug==2.3.7
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
cloudinary>=1.36.0
Flask-Mail==0.9.1
email-validator==2.1.0
